refactor: Split loader into seperate components
