+++ rustfmt.toml
@@ -1,2 +1,8 @@
 edition = "2021"
 reorder_imports = true
+#format_code_in_doc_comments = true
+#imports_granularity = "Item"
+#group_imports = "StdExternalCrate"
+#where_single_line = true
+#trailing_comma = "Vertical"
+#overflow_delimited_expr = true
