+++ src/aws/mod.rs
@@ -1,7 +1,9 @@
 //! AWS service signer
 //!
 //! Only sigv4 has been supported.
-pub mod loader;
 pub mod v4;
 
+mod config;
 mod constants;
+mod credential;
+mod region;
