+++ tests/aws/v4.rs
@@ -1,11 +1,15 @@
-use anyhow::Result;
-use http::{Request, StatusCode};
-use log::{debug, warn};
-use percent_encoding::{utf8_percent_encode, NON_ALPHANUMERIC};
-use reqsign::AwsV4Signer;
-use reqwest::blocking::Client;
 use std::env;
 use std::str::FromStr;
+
+use anyhow::Result;
+use http::Request;
+use http::StatusCode;
+use log::debug;
+use log::warn;
+use percent_encoding::utf8_percent_encode;
+use percent_encoding::NON_ALPHANUMERIC;
+use reqsign::AwsV4Signer;
+use reqwest::blocking::Client;
 use time::Duration;
 
 fn init_signer() -> Option<AwsV4Signer> {
