+++ tests/azure/storage.rs
@@ -1,11 +1,14 @@
-use anyhow::Result;
-use http::StatusCode;
-use log::{debug, warn};
-use percent_encoding::{utf8_percent_encode, NON_ALPHANUMERIC};
-use reqsign::azure::storage::Signer;
-use reqwest::blocking::Client;
 use std::env;
 use std::str::FromStr;
+
+use anyhow::Result;
+use http::StatusCode;
+use log::debug;
+use log::warn;
+use percent_encoding::utf8_percent_encode;
+use percent_encoding::NON_ALPHANUMERIC;
+use reqsign::azure::storage::Signer;
+use reqwest::blocking::Client;
 
 fn init_signer() -> Option<Signer> {
     let _ = env_logger::builder().is_test(true).try_init();
